Data of a replication study of C.R. Cox, J. Arndt, T. Pyszczynski, J. Greenberg, A. Abdollahi, S. Solomon (2008, JPSP, 94(4), Exp. 6)

This is the data of a replication study performed by J. Wissink, G. Hoogendoorn, H. Brohmer, M. Verschoor, J. Krijnen, and M. Zeelenberg as part of the Reproducibility Project: Psychology. The target result of this replication was the finding in Experiment 6 of Cox et al. (2008) that participants who scored low on avoidance but high on anxiety demonstrated an increased relative preference for a parent after mortality salience as opposed to dental pain, b = -32.04, SE = 14.47, t = -2.22, p = .03.

cox

Format

A data frame with 200 rows and 67 variables.

ID

Participant identifier

sex

The participant's sex

age

The participant's age

condition

The experimental condition: mortality salience or dental pain

avoidance

Attachment avoidance score as assessed with the Relationship Scales Questionnaire

anxiety

Attachment anxiety score as assessed with the Relationship Scales Questionnaire

affect_positive

Sum of positive PANAS items

affect_negative

Sum of negative PANAS items

call_parent

Minutes allocated (out of 100) to call a parent

call_siblings

Minutes allocated (out of 100) to call a sibling

call_partner

Minutes allocated (out of 100) to call a romantic partner

call_friend

Minutes allocated (out of 100) to call a close friend

Source

https://osf.io/5tbxf/

Details

For more information on the Reproducibility Project: Psychology, see https://osf.io/ezcuj/. The individual scores on the PANAS and RSQ items are not included in this dataset.

References

Cox, C.R., Arndt, J., Pyszczynski, T., Greenberg, J., Abdollahi, A., & Solomon, S. (2008) Terror management and adults’ attachment to their parents: The safe haven remains. Journal of Personality and Social Psychology, 94(4), 696-717, https://dx.doi.org/10.1037/0022-3514.94.4.696)