Tidy the output of a statistics object

tidy_stats is used to convert the output of a statistical object to a list of organized statistics. The tidy_stats function is automatically run when add_stats is used, so there is generally no need to use this function explicitly. It can be used, however, to peek at how the output of a specific analysis will be organized.

Usage

# S3 method for BFBayesFactor
tidy_stats(x, args = NULL)

# S3 method for rcorr
tidy_stats(x, args = NULL)

tidy_stats(x, args = NULL)

# S3 method for htest
tidy_stats(x, args = NULL)

# S3 method for pairwise.htest
tidy_stats(x, args = NULL)

# S3 method for lm
tidy_stats(x, args = NULL)

# S3 method for glm
tidy_stats(x, args = NULL)

# S3 method for anova
tidy_stats(x, args = NULL)

# S3 method for aov
tidy_stats(x, args = NULL)

# S3 method for aovlist
tidy_stats(x, args = NULL)

# S3 method for confint
tidy_stats(x, args = NULL)

# S3 method for afex_aov
tidy_stats(x, args = NULL)

# S3 method for mixed
tidy_stats(x, args = NULL)

# S3 method for brmsfit
tidy_stats(x, args = NULL)

# S3 method for effectsize_difference
tidy_stats(x, args = NULL)

# S3 method for effsize
tidy_stats(x, args = NULL)

# S3 method for emmGrid
tidy_stats(x, args = NULL)

# S3 method for summary_emm
tidy_stats(x, args = NULL)

# S3 method for emm_list
tidy_stats(x, args = NULL)

# S3 method for icclist
tidy_stats(x, args = NULL)

# S3 method for lavaan
tidy_stats(x, args = NULL)

# S3 method for lmerMod
tidy_stats(x, args = NULL)

# S3 method for lmerModLmerTest
tidy_stats(x, args = NULL)

# S3 method for predictions
tidy_stats(x, args = NULL)

# S3 method for comparisons
tidy_stats(x, args = NULL)

# S3 method for rma.uni
tidy_stats(x, args = NULL)

# S3 method for rma.mh
tidy_stats(x, args = NULL)

# S3 method for rma.peto
tidy_stats(x, args = NULL)

# S3 method for rma.glmm
tidy_stats(x, args = NULL)

# S3 method for rma.mv
tidy_stats(x, args = NULL)

# S3 method for confint.rma
tidy_stats(x, args = NULL)

# S3 method for list.confint.rma
tidy_stats(x, args = NULL)

# S3 method for anova.rma
tidy_stats(x, args = NULL)

# S3 method for permutest.rma.uni
tidy_stats(x, args = NULL)

# S3 method for tes
tidy_stats(x, args = NULL)

# S3 method for matreg
tidy_stats(x, args = NULL)

# S3 method for ranktest
tidy_stats(x, args = NULL)

# S3 method for regtest
tidy_stats(x, args = NULL)

# S3 method for fsn
tidy_stats(x, args = NULL)

# S3 method for hc.rma.uni
tidy_stats(x, args = NULL)

# S3 method for list.rma
tidy_stats(x, args = NULL)

# S3 method for lme
tidy_stats(x, args = NULL)

# S3 method for nlme
tidy_stats(x, args = NULL)

# S3 method for anova.lme
tidy_stats(x, args = NULL)

# S3 method for gls
tidy_stats(x, args = NULL)

# S3 method for psych
tidy_stats(x, args = NULL)

# S3 method for tidystats
tidy_stats(x, args = NULL)

# S3 method for tidystats_descriptives
tidy_stats(x, args = NULL)

# S3 method for tidystats_counts
tidy_stats(x, args = NULL)

Arguments

x

The output of a statistical test.

Details

Please note that not all statistical tests are supported. See 'Details' below for a list of supported statistical tests.

Currently supported functions:

stats:

t.test()

cor.test()

chisq.test()

wilcox.test()

fisher.test()

oneway.test()

lm()

glm()

aov()

anova()

lme4/lmerTest:

lmer()

BayesFactor:

generalTestBF()

lmBF()

regressionBF()

ttestBF()

anovaBF()

correlationBF()

contingencyTableBF()

proportionBF()

meta.ttestBF()

tidystats:

describe_data()

count_data()

Methods (by class)

tidy_stats(BFBayesFactor): tidy_stats method for class 'BayesFactor'

tidy_stats(rcorr): tidy_stats method for class 'rcorr'

tidy_stats(htest): tidy_stats method for class 'htest'

tidy_stats(pairwise.htest): tidy_stats method for class 'pairwise.htest'

tidy_stats(lm): tidy_stats method for class 'lm'

tidy_stats(glm): tidy_stats method for class 'glm'

tidy_stats(anova): tidy_stats method for class 'anova'

tidy_stats(aov): tidy_stats method for class 'aov'

tidy_stats(aovlist): tidy_stats method for class 'aovlist'

tidy_stats(confint): tidy_stats method for class 'confint'

tidy_stats(afex_aov): tidy_stats method for class 'afex_aov'

tidy_stats(mixed): tidy_stats method for class 'afex_aov'

tidy_stats(brmsfit): tidy_stats method for class 'brmsfit'

tidy_stats(effectsize_difference): tidy_stats method for class 'effectsize_difference'

tidy_stats(effsize): tidy_stats method for class 'effsize'

tidy_stats(emmGrid): tidy_stats method for class 'emmGrid'

tidy_stats(summary_emm): tidy_stats method for class 'summary_emm'

tidy_stats(emm_list): tidy_stats method for class 'emm_list'

tidy_stats(icclist): tidy_stats method for class 'icclist'

tidy_stats(lavaan): tidy_stats method for class 'lavaan'

tidy_stats(lmerMod): tidy_stats method for class 'lmerMod'

tidy_stats(lmerModLmerTest): tidy_stats method for class 'lmerModLmerTest'

tidy_stats(predictions): tidy_stats method for class 'predictions'

tidy_stats(comparisons): tidy_stats method for class 'comparisons'

tidy_stats(rma.uni): tidy_stats method for class 'rma.uni'

tidy_stats(rma.mh): tidy_stats method for class 'rma.mh'

tidy_stats(rma.peto): tidy_stats method for class 'rma.peto'

tidy_stats(rma.glmm): tidy_stats method for class 'rma.glmm'

tidy_stats(rma.mv): tidy_stats method for class 'rma.mv'

tidy_stats(confint.rma): tidy_stats method for class 'confint.rma'

tidy_stats(list.confint.rma): tidy_stats method for class 'list.confint.rma'

tidy_stats(anova.rma): tidy_stats method for class 'anova.rma'

tidy_stats(permutest.rma.uni): tidy_stats method for class 'permutest.rma.uni'

tidy_stats(tes): tidy_stats method for class 'tes'

tidy_stats(matreg): tidy_stats method for class 'matreg'

tidy_stats(ranktest): tidy_stats method for class 'ranktest'

tidy_stats(regtest): tidy_stats method for class 'regtest'

tidy_stats(fsn): tidy_stats method for class 'fsn'

tidy_stats(hc.rma.uni): tidy_stats method for class 'hc.rma.uni'

tidy_stats(list.rma): tidy_stats method for class 'list.rma'

tidy_stats(lme): tidy_stats method for class 'lme'

tidy_stats(nlme): tidy_stats method for class 'nlme'

tidy_stats(anova.lme): tidy_stats method for class 'anova.lme'

tidy_stats(gls): tidy_stats method for class 'gls'

tidy_stats(psych): tidy_stats method for class 'psych'

tidy_stats(tidystats): tidy_stats method for class 'tidystats'

tidy_stats(tidystats_descriptives): tidy_stats method for class 'tidystats_descriptives'

tidy_stats(tidystats_counts): tidy_stats method for class 'tidystats_counts'

Examples

# Conduct statistical tests
# t-test:
sleep_test <- t.test(extra ~ group, data = sleep, paired = TRUE)

# lm:
ctl <- c(4.17, 5.58, 5.18, 6.11, 4.50, 4.61, 5.17, 4.53, 5.33, 5.14)
trt <- c(4.81, 4.17, 4.41, 3.59, 5.87, 3.83, 6.03, 4.89, 4.32, 4.69)
group <- gl(2, 10, 20, labels = c("Ctl", "Trt"))
weight <- c(ctl, trt)
lm_D9 <- lm(weight ~ group)

# ANOVA:
npk_aov <- aov(yield ~ block + N * P * K, npk)

# Tidy the statistics and store each analysis in a separate variable
list_sleep_test <- tidy_stats(sleep_test)
#> Error in UseMethod("tidy_stats"): no applicable method for 'tidy_stats' applied to an object of class "htest"
list_lm_D9 <- tidy_stats(lm_D9)
#> Error in UseMethod("tidy_stats"): no applicable method for 'tidy_stats' applied to an object of class "lm"
list_npk_aov <- tidy_stats(npk_aov)
#> Error in UseMethod("tidy_stats"): no applicable method for 'tidy_stats' applied to an object of class "c('aov', 'lm')"

# Now you can inspect each of these variables, e.g.,:
names(list_sleep_test)
#> Error in eval(expr, envir, enclos): object 'list_sleep_test' not found
str(list_sleep_test)
#> Error in eval(expr, envir, enclos): object 'list_sleep_test' not found