Tidy the output of a statistics object

tidy_stats is used to convert the output of a statistical object to a list of organized statistics. This output can then be added to a list using the add_stats function of this package. The tidy_stats function is automatically run when add_stats is used, so there is generally no need to use this function explicitly. It can be used, however, to quickly peek at how the output of a specific analysis will be organized.Please note that not all statistical tests are supported. See 'Details' below for a list of supported statistical tests.

tidy_stats(x)

# S3 method for htest
tidy_stats(x)

# S3 method for lm
tidy_stats(x)

# S3 method for glm
tidy_stats(x)

# S3 method for anova
tidy_stats(x)

# S3 method for aov
tidy_stats(x)

# S3 method for aovlist
tidy_stats(x)

# S3 method for tidystats_descriptives
tidy_stats(x)

# S3 method for tidystats_counts
tidy_stats(x)

# S3 method for lmerMod
tidy_stats(x)

# S3 method for lmerModLmerTest
tidy_stats(x)

# S3 method for BFBayesFactor
tidy_stats(x)

# S3 method for afex_aov
tidy_stats(x)

Arguments

x
The output of a statistical test.

Details

Currently supported functions:

t.test()

cor.test()

chisq.test()

wilcox.test()

fisher.test()

oneway.test()

aov()

lm()

Methods (by class)

htest: tidy_stats method for class 'htest'

lm: tidy_stats method for class 'lm'

glm: tidy_stats method for class 'glm'

anova: tidy_stats method for class 'anova'

aov: tidy_stats method for class 'aov'

aovlist: tidy_stats method for class 'aovlist'

tidystats_descriptives: tidy_stats method for class 'tidystats_descriptives'

tidystats_counts: tidy_stats method for class 'tidystats_counts'

lmerMod: tidy_stats method for class 'lmerMod'

lmerModLmerTest: tidy_stats method for class 'lmerModLmerTest'

BFBayesFactor: tidy_stats method for class 'BayesFactor'

afex_aov: tidy_stats method for class 'afex_aov'

Examples

# Conduct statistical tests
# t-test:
sleep_test <- t.test(extra ~ group, data = sleep, paired = TRUE)

# lm:
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm_D9 <- lm(weight ~ group)

# ANOVA:
npk_aov <- aov(yield ~ block + N*P*K, npk)

# Tidy the statistics and store each analysis in a separate variable
list_sleep_test <- tidy_stats(sleep_test)
list_lm_D9 <- tidy_stats(lm_D9)
list_npk_aov <- tidy_stats(npk_aov)

# Now you can inspect each of these variables, e.g.,:
names(list_sleep_test)

#> [1] "method"      "var_equal"   "name"        "statistics"  "alternative"
#> [6] "package"    

str(list_sleep_test)

#> List of 6
#>  $ method     : chr "Paired t-test"
#>  $ var_equal  : logi TRUE
#>  $ name       : chr "extra by group"
#>  $ statistics :List of 6
#>   ..$ estimate :List of 2
#>   .. ..$ name : chr "mean difference"
#>   .. ..$ value: num -1.58
#>   ..$ SE       : num 0.389
#>   ..$ statistic:List of 2
#>   .. ..$ name : chr "t"
#>   .. ..$ value: num -4.06
#>   ..$ df       : num 9
#>   ..$ p        : num 0.00283
#>   ..$ CI       :List of 3
#>   .. ..$ CI_level: num 0.95
#>   .. ..$ CI_lower: num -2.46
#>   .. ..$ CI_upper: num -0.7
#>  $ alternative:List of 2
#>   ..$ direction : chr "two.sided"
#>   ..$ null_value: num 0
#>  $ package    :List of 2
#>   ..$ name   : chr "stats"
#>   ..$ version: chr "4.0.2"