Convert a tidystats list to a data frame

tidy_stats_to_data_frame() converts a tidystats list to a data frame, which can then be used to extract specific statistics using standard subsetting functions (e.g., dplyr::filter()).

Usage

tidy_stats_to_data_frame(x)

Arguments

x

A tidystats list.

Examples

# Conduct statistical tests
sleep_test <- t.test(extra ~ group, data = sleep, paired = TRUE)

ctl <- c(4.17, 5.58, 5.18, 6.11, 4.50, 4.61, 5.17, 4.53, 5.33, 5.14)
trt <- c(4.81, 4.17, 4.41, 3.59, 5.87, 3.83, 6.03, 4.89, 4.32, 4.69)
group <- gl(2, 10, 20, labels = c("Ctl", "Trt"))
weight <- c(ctl, trt)
lm_D9 <- lm(weight ~ group)

npk_aov <- aov(yield ~ block + N * P * K, npk)

# Create an empty list
statistics <- list()

# Add statistics
statistics <- statistics |>
  add_stats(sleep_test) |>
  add_stats(lm_D9, type = "primary", preregistered = TRUE) |>
  add_stats(npk_aov, notes = "An ANOVA example")

# Convert the list to a data frame
df <- tidy_stats_to_data_frame(statistics)

# Select all the p-values
filter(df, statistic == "p")
#> # A tibble: 11 × 11
#>    identifier group_1    group_2 statistic symbol subscript lower    value upper
#>    <chr>      <chr>      <chr>   <chr>     <chr>  <chr>     <dbl>    <dbl> <dbl>
#>  1 sleep_test NA         NA      p         NA     NA           NA 2.83e- 3    NA
#>  2 lm_D9      Model      NA      p         NA     NA           NA 2.49e- 1    NA
#>  3 lm_D9      Coefficie… (Inter… p         NA     NA           NA 9.55e-15    NA
#>  4 lm_D9      Coefficie… groupT… p         NA     NA           NA 2.49e- 1    NA
#>  5 npk_aov    Terms      block   p         NA     NA           NA 1.59e- 2    NA
#>  6 npk_aov    Terms      N       p         NA     NA           NA 4.37e- 3    NA
#>  7 npk_aov    Terms      P       p         NA     NA           NA 4.75e- 1    NA
#>  8 npk_aov    Terms      K       p         NA     NA           NA 2.88e- 2    NA
#>  9 npk_aov    Terms      N:P     p         NA     NA           NA 2.63e- 1    NA
#> 10 npk_aov    Terms      N:K     p         NA     NA           NA 1.69e- 1    NA
#> 11 npk_aov    Terms      P:K     p         NA     NA           NA 8.63e- 1    NA
#> # ℹ 2 more variables: interval <chr>, level <dbl>