Convert a tidystats list to a data frame

tidy_stats_to_data_frame converts a tidystats list to a data frame, which can then be used to easily extract specific statistics using standard subsetting functions (e.g., dplyr::filter).

tidy_stats_to_data_frame(x)

Arguments

x
A tidystats list.

Examples

# Load dplyr for access to the piping operator
library(dplyr)

# Conduct statistical tests
t_test_1 <- t.test(1:10, y = c(7:20))
t_test_2 <- t.test(1:10, y = c(7:20, 200))
t_test_3 <- t.test(extra ~ group, data = sleep)

#' # Create an empty list
results <- list()

# Add tests to the empty list
results <- results %>%
  add_stats(t_test_1) %>%
  add_stats(t_test_2) %>%
  add_stats(t_test_3)

# Convert the list to a data frame
results_df <- tidy_stats_to_data_frame(results)

# Select all the p-values
results_df %>%
  filter(statistic == "p") %>%
  pull(value)

#> [1] 1.855282e-05 1.245135e-01 7.939414e-02